alphabot-back/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
psycopg[binary]>=3.1.18
pydantic>=2.6.4
pydantic-settings>=2.2.1

##토큰(JWT)을 통해 사용자를 식별하는 의존성(Dependency) 구현을 위한 추가
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart